Fix food spawn issue

+++ src/SnakeMeECS/Assets/ProjectCore/Features/Map/Components/MapProperties.cs
@@ -10,6 +10,6 @@ namespace ProjectCore.Features.Map.Components
 
         public float DistanceBetweenCells;
         
-        public float MapHeight;
+        public int MapHeight;
     }
 }
